Implement callbacks for RAUW and Delete in the `LLVMContext`
Allows registering custom callback functions to be called when a value is RAUW'd or deleted. Uses a RAII-class with exclusive ownership semantics of these callbacks to register and deregister them from the context automatically.

The implementation is not fully finished, see the TODO comments in this commit diff.

The callbacks seem to be zero-cost when they're not used, meaning they have no measurable effect on compile-time in this case, as they are guarded behind a check, although they do require the RAUW and Delete implementations to load the LLVMContextImpl, which was not the case before. But as this has no dependence on the rest of the code, it seems that out-of-order execution takes care of not slowing down the critical path.